Create Book Update
I was provided a list from a client of "books" in their database which needed to be removed but could not via a front-end solution. I created the script to seamlessly move them to a new book (also created in the script) and drop the ones they no longer needed.