Create Rebuild mySQL
A rebuild was required which involved a large number of tables and cleanup procedures. During the process, I ran into a lot of system limitations as the fdktickets table had 10+ million entries. To solve the issue, I created procedures and various states of the tables having or not having indices.